backend/migrations/versions/b2c3d4e5f6a7_milestone16_add_installed_at.py
"""milestone16_add_installed_at

Revision ID: b2c3d4e5f6a7
Revises: a1b2c3d4e5f6
Create Date: 2026-03-13 17:00:00.000000

"""
from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision = 'b2c3d4e5f6a7'
down_revision = 'a1b2c3d4e5f6'
branch_labels = None
depends_on = None


def upgrade():
    op.add_column('instances', sa.Column('installed_at', sa.DateTime(), nullable=True))


def downgrade():
    op.drop_column('instances', 'installed_at')
