backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
requests==2.32.3
pyotp==2.9.0
gunicorn==23.0.0
